+++ npm-packages/shared-utils/src/types/types.ts
@@ -116,7 +116,6 @@ export enum ExtensionTypeEnum {
 export type ExtensionConfig = {
   id: string;
   version: string;
-  mfVersion: string;
   author?: string;
   displayName?: string;
   description?: string;