+++ web/components/explorer/project/project-item.tsx
@@ -1,40 +1,11 @@
-"use client";
-
-import { useContext, useState } from "react";
-import { EditorContext } from "../providers/editor-context-provider";
-import { Button } from "@heroui/react";
 import { usePlatformApi } from "@/lib/hooks/use-platform-api";
 import { ContextMenuState, ProjectInfo } from "@/lib/types";
-import ContextMenu from "../interface/context-menu";
-import ProjectSettingsModal from "../modals/project-settings-modal";
+import { Button } from "@heroui/react";
+import { useContext, useState } from "react";
+import ContextMenu from "../../interface/context-menu";
+import { EditorContext } from "../../providers/editor-context-provider";
 
-export default function ProjectList() {
-  const editorContext = useContext(EditorContext);
-  const [settingsOpen, setSettingsOpen] = useState(false);
-  const [settingsProject, setSettingsProject] = useState<
-    ProjectInfo | undefined
-  >(undefined);
-
-  return (
-    <div className="flex w-full flex-col gap-2">
-      {editorContext?.editorStates.projectsInfo?.map((project, index) => (
-        <ProjectTab
-          key={index}
-          project={project}
-          setSettingsOpen={setSettingsOpen}
-          setSettingsProject={setSettingsProject}
-        />
-      ))}
-      <ProjectSettingsModal
-        isOpen={settingsOpen}
-        setIsOpen={setSettingsOpen}
-        projectInfo={settingsProject}
-      />
-    </div>
-  );
-}
-
-function ProjectTab({
+export default function ProjectItem({
   project,
   setSettingsOpen,
   setSettingsProject,
@@ -51,24 +22,36 @@ function ProjectTab({
     isOpen: false,
   });
 
-  function openProject(projectName: string) {
-    const uri =
-      editorContext?.persistSettings?.projectHomePath + "/" + projectName;
+  const projectName = project.name;
+  const projectCtime = project.ctime
+    ? formatDateTime(project.ctime)
+    : "Unknown";
 
-    platformApi
-      ?.listPathContent(uri, {
-        include: "all",
-        isRecursive: true,
-      })
-      .then((objects) => {
-        editorContext?.setEditorStates((prev) => {
-          return {
-            ...prev,
-            project: projectName,
-            projectContent: objects,
-          };
-        });
-      });
+  function openProject(projectName: string) {
+    editorContext?.setEditorStates((prev) => {
+      return {
+        ...prev,
+        project: projectName,
+      };
+    });
+
+    // TODO: move this to when workspace is loaded
+    // const uri =
+    //   editorContext?.persistSettings?.projectHomePath + "/" + projectName;
+    // platformApi
+    //   ?.listPathContent(uri, {
+    //     include: "all",
+    //     isRecursive: true,
+    //   })
+    //   .then((objects) => {
+    //     editorContext?.setEditorStates((prev) => {
+    //       return {
+    //         ...prev,
+    //         project: projectName,
+    //         projectContent: objects,
+    //       };
+    //     });
+    //   });
   }
 
   function formatDateTime(date: Date) {
@@ -89,7 +72,7 @@ function ProjectTab({
         onPress={(e) => {
           // Only open project if context menu is not open
           if (!contextMenuState.isOpen) {
-            openProject(project.name);
+            openProject(projectName);
           }
         }}
         onContextMenu={(e) => {
@@ -107,10 +90,8 @@ function ProjectTab({
         }}
       >
         <div className="flex w-full flex-col items-start justify-center">
-          <p>{project.name}</p>
-          <p className="text-xs">
-            {"Created: " + formatDateTime(project.ctime)}
-          </p>
+          <p>{projectName}</p>
+          <p className="text-xs">{"Created: " + projectCtime}</p>
         </div>
       </Button>
       <ContextMenu state={contextMenuState} setState={setContextMenuState}>