+++ web/components/interface/navigation/nav-side-menu.tsx
@@ -1,7 +1,16 @@
-import { AnimatePresence, motion } from "framer-motion";
-import { useMediaQuery } from "react-responsive";
-import Explorer from "../../explorer/explorer";
+import AppExplorer from "@/components/explorer/app/app-explorer";
+import ProjectExplorer from "@/components/explorer/project/project-explorer";
+import Tabs from "@/components/misc/tabs";
+import ProjectSettingsModal from "@/components/modals/project-settings-modal";
+import { EditorContext } from "@/components/providers/editor-context-provider";
+import useExplorer from "@/lib/hooks/use-explorer";
+import { isWeb } from "@/lib/platform-api/platform-checker";
+import { TabItem } from "@/lib/types";
 import { Button } from "@heroui/react";
+import { AnimatePresence, motion } from "framer-motion";
+import { useContext, useState } from "react";
+import { useMediaQuery } from "react-responsive";
+import FileSystemExplorer from "../../explorer/file-system/fs-explorer";
 import Icon from "../../misc/icon";
 
 export default function NavSideMenu({
@@ -35,7 +44,8 @@ export default function NavSideMenu({
                   <Icon name="arrow_back" />
                 </Button>
               </div>
-              <Explorer setIsMenuOpen={setIsMenuOpen} />
+
+              <PanelContent setIsMenuOpen={setIsMenuOpen} />
             </div>
           </div>
         </MenuPanel>
@@ -89,5 +99,98 @@ function MenuPanel({ children }: { children?: React.ReactNode }) {
         </motion.div>
       )}
     </>
+  );
+}
+
+function PanelContent({
+  setIsMenuOpen,
+}: {
+  setIsMenuOpen: (isOpen: boolean) => void;
+}) {
+  const editorContext = useContext(EditorContext);
+
+  const { selectAndSetProjectHome } = useExplorer();
+
+  const [isProjectSettingsModalOpen, setIsProjectSettingsModalOpen] =
+    useState(false);
+
+  const tabItems: TabItem[] = [
+    {
+      name: "Apps",
+      description: "List of apps",
+      icon: "apps",
+    },
+    {
+      name: "Workspace",
+      description: "Project workspace",
+      icon: "folder",
+    },
+  ];
+  const [selectedTabIndex, setSelectedTabIndex] = useState<number>(0);
+
+  // Choose project home path
+  if (!isWeb() && !editorContext?.persistSettings?.projectHomePath) {
+    return (
+      <div className="bg-content2 h-full w-full space-y-2 p-4">
+        <p>
+          You have not set a project home path yet. Please set a project home
+          path to continue. All your projects will be saved in this directory.
+        </p>
+        <Button
+          className="w-full"
+          onPress={() => {
+            selectAndSetProjectHome();
+          }}
+        >
+          Select Project Home Path
+        </Button>
+      </div>
+    );
+  }
+  // Pick project if no project is opened
+  else if (!editorContext?.editorStates.project) {
+    return (
+      <div className="bg-content2 h-full w-full space-y-2 overflow-y-auto px-4">
+        <p className="text-center text-lg font-medium">View Projects</p>
+        <Button
+          className="w-full"
+          onPress={() => {
+            setIsProjectSettingsModalOpen(true);
+          }}
+        >
+          New Project
+        </Button>
+        <ProjectExplorer />
+        <ProjectSettingsModal
+          isOpen={isProjectSettingsModalOpen}
+          setIsOpen={setIsProjectSettingsModalOpen}
+        />
+      </div>
+    );
+  }
+
+  return (
+    <div className="relative h-full w-full px-4">
+      <div className="flex w-full justify-center">
+        <div className="w-fit">
+          <Tabs
+            tabItems={tabItems}
+            selectedItem={tabItems[selectedTabIndex]}
+            setSelectedItem={(item) => {
+              const index = tabItems.findIndex(
+                (tab) => tab.name === item?.name,
+              );
+              setSelectedTabIndex(index !== -1 ? index : 0);
+            }}
+            isClosable={false}
+          />
+        </div>
+      </div>
+      {tabItems[selectedTabIndex]?.name === "Apps" ? (
+        <AppExplorer />
+      ) : (
+        <FileSystemExplorer setIsMenuOpen={setIsMenuOpen} />
+      )}
+    </div>
   );
 }