+++ web/lib/hooks/use-tab-view-manager.ts
@@ -1,6 +1,8 @@
-import { useCallback, useContext, useEffect, useState } from "react";
 import { EditorContext } from "@/components/providers/editor-context-provider";
+import { IMCContext } from "@/components/providers/imc-provider";
 import { ViewModeEnum } from "@pulse-editor/shared-utils";
+import { useContext, useEffect, useState } from "react";
+import { v4 } from "uuid";
 import {
   AppViewConfig,
   CanvasViewConfig,
@@ -8,7 +10,6 @@ import {
   MenuAction,
   TabView,
 } from "../types";
-import { IMCContext } from "@/components/providers/imc-provider";
 import { useMenuActions } from "./use-menu-actions";
 
 export function useTabViewManager() {
@@ -47,7 +48,7 @@ export function useTabViewManager() {
       name: "Close Workflow",
       actionFunc: async () => {
         if (activeTabView) {
-          closeView(activeTabView);
+          closeTabView(activeTabView);
         }
       },
       menuCategory: "file",
@@ -182,7 +183,7 @@ export function useTabViewManager() {
     return editorContext.persistSettings?.defaultFileTypeExtensionMap?.[suffix];
   }
 
-  function closeView(view: TabView) {
+  function closeTabView(view: TabView) {
     if (!editorContext) {
       throw new Error("Editor context is not available");
     }
@@ -208,7 +209,7 @@ export function useTabViewManager() {
   /**
    * Clear all views
    */
-  function closeAllViews() {
+  function closeAllTabViews() {
     if (!editorContext) {
       throw new Error("Editor context is not available");
     }
@@ -238,17 +239,15 @@ export function useTabViewManager() {
       throw new Error("IMC context is not available");
     }
 
+    const newTabView: TabView = {
+      type,
+      config,
+    };
+
     editorContext.setEditorStates((prev) => {
       return {
         ...prev,
-        tabViews: [
-          ...prev.tabViews,
-          {
-            viewId: config.viewId,
-            type,
-            config,
-          },
-        ],
+        tabViews: [...prev.tabViews, newTabView],
         tabIndex: prev.tabViews.length,
       };
     });
@@ -257,22 +256,29 @@ export function useTabViewManager() {
       // Wait for app view to be initialized. This is because unlike canvas views,
       // app views need to load the app first before it can render the view.
       await imcContext.resolveWhenViewInitialized(config.viewId);
-    }
-    else if (type === ViewModeEnum.Canvas) { 
+    } else if (type === ViewModeEnum.Canvas) {
       // Open explorer for canvas views
       editorContext.setEditorStates((prev) => ({
         ...prev,
-        isSideMenuOpen: true
-      }))
+        isSideMenuOpen: true,
+      }));
     }
+
+    return newTabView;
   }
 
   async function createAppViewInCanvasView(appConfig: AppViewConfig) {
     if (!editorContext) {
       throw new Error("Editor context is not available");
     }
-    const currentTab =
-      editorContext.editorStates.tabViews[editorContext.editorStates.tabIndex];
+    let currentTab = activeTabView;
+
+    if (!currentTab) {
+      currentTab = await createTabView(ViewModeEnum.Canvas, {
+        viewId: `canvas-${v4()}`,
+      } as CanvasViewConfig);
+    }
+
     if (currentTab?.type !== ViewModeEnum.Canvas) {
       throw new Error("Current tab is not a canvas");
     }
@@ -296,6 +302,35 @@ export function useTabViewManager() {
     });
 
     await imcContext?.resolveWhenViewInitialized(appConfig.viewId);
+  }
+
+  async function deleteAppViewInCanvasView(viewId: string) {
+    if (!editorContext) {
+      throw new Error("Editor context is not available");
+    }
+    const currentTab = activeTabView;
+    if (currentTab?.type !== ViewModeEnum.Canvas) {
+      throw new Error("Current tab is not a canvas");
+    }
+
+    const newCanvasConfig: CanvasViewConfig = {
+      ...currentTab.config,
+      nodes: (currentTab.config as CanvasViewConfig).nodes?.filter(
+        (config) => config.viewId !== viewId,
+      ),
+    };
+
+    editorContext.setEditorStates((prev) => {
+      const newViews = [...prev.tabViews];
+      newViews[prev.tabIndex] = {
+        ...currentTab,
+        config: newCanvasConfig,
+      };
+      return {
+        ...prev,
+        tabViews: newViews,
+      };
+    });
   }
 
   function getId(nameOrUrl: string) {
@@ -358,10 +393,11 @@ export function useTabViewManager() {
     selectTab,
     viewCount,
     openFileInView,
-    closeView,
-    closeAllViews,
+    closeTabView,
+    closeAllTabViews,
     createTabView,
     createAppViewInCanvasView,
+    deleteAppViewInCanvasView,
     findAppInTabView,
   };
 }