+++ npm-packages/react-api/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pulse-editor/react-api
 
+## 0.1.1-alpha.34
+
+### Patch Changes
+
+- Revert mfversion in extension config
+- Updated dependencies
+  - @pulse-editor/shared-utils@0.1.1-alpha.34
+
 ## 0.1.1-alpha.33
 
 ### Patch Changes